Use HostnameEndpoint in client for both IPv4 and IPv6
This change was originally introduced and then deleted.
Now is reintroduced after tests on both protocols.
HostnameEndpoint is needed in some configurations to resolve
IPv6-only hostnames, so it is required. The connection now works
with both protocols on clients and servers.

NOTE: now a wrong server/address/port DOES NOT trigger a Connection
with Server Failed error immediately, the error stays in Deferred.
This must be fixed before putting this code in production.