Release some files under the MIT license.
The freedesktop project requires that the AppStream metadata are
released under a subset of licenses that does not include Apache 2.0.
For this reason, we decided to change the license of the README file
and release it under the MIT license (after obtaining explicit
permission from all the contributors to that file). With the same
goal, the script that generates the AppImage and the AppStreams
metadata is also released under the MIT license. The README file now
clearly states the project license and that some files are released
under a different license, as declared in their header or in our
third-party notices file.