Fix running IINA from frozen app
Apparently, `iina-cli` gets confused when launched from a frozen app
and automatically adds `--stdin` to its passed launch arguments.
But then, it waits for a file to be piped and, because there is none,
the player crashes almost immediately. Sending `--no-stdin` to the
process resolves the ambiguity and does not cause any harm if
Syncplay is started from sources.